Refactor: Added history folder and moved files inside
