+++ app/src/components/market/sections/common_sections/outcome_table/history/history_section/index.tsx
@@ -3,10 +3,10 @@ import { BigNumber } from 'ethers/utils'
 import moment from 'moment'
 import React, { useEffect, useMemo, useState } from 'react'
 
-import { EARLIEST_MAINNET_BLOCK_TO_CHECK, EARLIEST_RINKEBY_BLOCK_TO_CHECK } from '../../../../common/constants'
-import { useConnectedWeb3Context, useMultipleQueries } from '../../../../hooks'
-import { isScalarMarket, keys, range, waitABit } from '../../../../util/tools'
-import { Period } from '../../../../util/types'
+import { EARLIEST_MAINNET_BLOCK_TO_CHECK, EARLIEST_RINKEBY_BLOCK_TO_CHECK } from '../../../../../../../common/constants'
+import { useConnectedWeb3Context, useMultipleQueries } from '../../../../../../../hooks'
+import { isScalarMarket, keys, range, waitABit } from '../../../../../../../util/tools'
+import { Period } from '../../../../../../../util/types'
 
 import { History_select } from './history_select'
